Fixes with Furkan
