+++ app/Components/SideBar.jsx
@@ -1,9 +1,9 @@
 import "../globals.css";
 import Link from "next/link";
 import Icons from "./Icons";
+import { useTranslations } from "next-intl";
 
 function SideBar() {
-  const t = useTranslations("SideBar")
   return (
     <>
       <div className="fixed top-7 left-6 z-50">
@@ -16,27 +16,27 @@ function SideBar() {
           <ul className="flex justify-around md:justify-evenly w-[85vw] list-none md:bg-transparent bg-black/75 backdrop-blur-xl p-[1em] rounded-[40px] md:flex-col md:items-center md:h-[25em] md:w-[7em] z-40">
             <li>
               <Link href="/home" className="mainLinks no-underline text-[#b1b1b1]">
-                <Icons.HomeIcon />
+                <Icons.HomeIcon t={useTranslations("SideBar")} />
               </Link>
             </li>
             <li>
               <Link href="/search" className="mainLinks no-underline text-[#b1b1b1]">
-                <Icons.SearchIcon />
+                <Icons.SearchIcon t={useTranslations("SideBar")} />
               </Link>
             </li>
             <li>
               <Link href="/courses/courseId" className="mainLinks no-underline text-[#b1b1b1]">
-                <Icons.CoursesIcon />
+                <Icons.CoursesIcon t={useTranslations("SideBar")} />
               </Link>
             </li>
             <li>
               <Link href="/saved" className="mainLinks no-underline text-[#b1b1b1]">
-                <Icons.SavedIcon />
+                <Icons.SavedIcon t={useTranslations("SideBar")} />
               </Link>
             </li>
             <li>
               <Link href="/profile" className="mainLinks no-underline text-[#b1b1b1]">
-                <Icons.ProfileIcon />
+                <Icons.ProfileIcon t={useTranslations("SideBar")} />
               </Link>
             </li>
           </ul>
