+++ app/[locale]/layout.jsx
@@ -2,6 +2,7 @@ import { useLocale } from "next-intl"
 import { notFound } from "next/navigation"
 import '../globals.css'
 import { Inter } from 'next/font/google'
+import {NextIntlClientProvider} from 'next-intl';
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -10,15 +11,24 @@ export const metadata = {
     description: 'LearnU is a platform for learning and teaching online.'
 }
 
-export default function LocaleLayout({ children, params }) {
+export default async function LocaleLayout({ children, params }) {
     const locale = useLocale()
     if (params.locale !== locale) {
         return notFound()
     }
+    let messages;
+    try {
+      messages = (await import(`../../messages/${locale}.json`)).default;
+    } catch (error) {
+      notFound();
+    }
+
     return (
         <html lang="{locale}">
             <body className={inter.className}>
-                {children}
+                <NextIntlClientProvider locale={locale} messages={messages}>
+                    {children}
+                </NextIntlClientProvider>
             </body>
         </html>
     )
