+++ messages/tr.json
@@ -15,7 +15,8 @@
             "HR": "İK",
             "Design": "Tasarım",
             "Astronomy": "Astronomi",
-            "Big Data": "Büyük Veri"
+            "Big Data": "Büyük Veri",
+            "More": "Daha Fazla Göster"
         }
     },
     "SideBar": {
